cleaned up tutorial code, added lots of comments
