+++ Tutorials/Amr/Advection_AmrCore/Source/AmrAdv.H
@@ -20,82 +20,135 @@ class AmrAdv
 public:
 
     ////////////////
-    // member functions
+    // public member functions
 
-    // in AmrAdv.cpp
+    // constructor - reads in parameters from inputs file
+    //             - sizes multilevel arrays and data structures
     AmrAdv ();
     virtual ~AmrAdv();
+
+    // advance solution to final time
+    void Evolve ();
+
+    // initializes multilevel data
+    void InitData ();
+
+    // Make a new level from scratch using provided BoxArray and DistributionMapping.
+    // Only used during initialization.
+    // overrides the pure virtual function in AmrCore
     virtual void MakeNewLevelFromCoarse (int lev, amrex::Real time, const amrex::BoxArray& ba,
 					 const amrex::DistributionMapping& dm) override;
+
+    // Remake an existing level using provided BoxArray and DistributionMapping and 
+    // fill with existing fine and coarse data.
+    // overrides the pure virtual function in AmrCore
     virtual void RemakeLevel (int lev, amrex::Real time, const amrex::BoxArray& ba,
 			      const amrex::DistributionMapping& dm) override;
+
+    // Delete level data
+    // overrides the pure virtual function in AmrCore
     virtual void ClearLevel (int lev) override;
 
-    // in AmrAdvInit.cpp
-    void InitData ();
+    // initialize data using a fortran routine to compute initial state
+    // overrides the pure virtual function in AmrCore
     virtual void MakeNewLevelFromScratch (int lev, amrex::Real time, const amrex::BoxArray& ba,
 					  const amrex::DistributionMapping& dm) override;
 
-    // in AmrAdvError.cpp
+    // tag all cells for refinement
+    // overrides the pure virtual function in AmrCore
     virtual void ErrorEst (int lev, amrex::TagBoxArray& tags, amrex::Real time, int ngrow) override;
-
-    // in AmrAdvEvolve.cpp
-    void Evolve ();
 
 private:
 
     ////////////////
-    // member functions
+    // private member functions
 
-    // in AmrAdv.cpp
+    // read in some parameters from inputs file
     void ReadParameters();
+
+    // set covered coarse cells to be the average of overlying fine cells
     void AverageDown ();
+
+    // more flexible version of AverageDown() that lets you average down across multiple levels
     void AverageDownTo (int crse_lev);
+
+    // compute the number of cells at a level
     long CountCells (int lev);
+
+    // compute a new multifab by coping in phi from valid region and filling ghost cells
+    // works for single level and 2-level cases (fill fine grid ghost by interpolating from coarse)
     void FillPatch (int lev, amrex::Real time, amrex::MultiFab& mf, int icomp, int ncomp);
+
+    // fill an entire multifab by interpolating from the coarser level
+    // this comes into play when a new level of refinement appears
     void FillCoarsePatch (int lev, amrex::Real time, amrex::MultiFab& mf, int icomp, int ncomp);
+
+    // utility to copy in data from phi_old and/or phi_new into another multifab
     void GetData (int lev, amrex::Real time, amrex::Array<amrex::MultiFab*>& data,
                   amrex::Array<amrex::Real>& datatime);
 
     // in AmrAdvEvolve.cpp
     void timeStep (int lev, amrex::Real time, int iteration);
+
+    // advance a single level for a single time step, updates flux registers
     void Advance (int lev, amrex::Real time, amrex::Real dt, int iteration, int ncycle);
+
+    // a wrapper for EstTimeStep(0
     void ComputeDt ();
+
+    // compute dt from CFL considerations
     Real EstTimeStep (int lev, bool local=false) const;
 
-    // in AmrAdvIO.cpp
+    // get plotfile name
     std::string PlotFileName (int lev) const;
+
+    // put together an array of multifabs for writing
     amrex::Array<const amrex::MultiFab*> PlotFileMF () const;
+
+    // set plotfile variables names
     amrex::Array<std::string> PlotFileVarNames () const;
+
+    // write plotfile to disk
     void WritePlotFile () const;
 
     ////////////////
-    // data members
+    // private data members
 
     amrex::Array<int> istep;      // which step?
     amrex::Array<int> nsubsteps;  // how many substeps on each level?
 
-    amrex::Array<amrex::Real> t_new;
+    // keep track of old time, new time, and time step at each level
+    amrex::Array<amrex::Real> t_new;  
     amrex::Array<amrex::Real> t_old;
     amrex::Array<amrex::Real> dt;
 
+    // array of multifabs to store the solution at each level of refinement
+    // after advancing a level we use "swap" to change pointers to old/new
+    // to avoid a copy
     amrex::Array<std::unique_ptr<amrex::MultiFab> > phi_new;
     amrex::Array<std::unique_ptr<amrex::MultiFab> > phi_old;
 
     // stores fluxes at coarse-fine interface for synchronization
     amrex::Array<std::unique_ptr<amrex::FluxRegister> > flux_reg; 
     
+    ////////////////
     // runtime parameters
 
-    int max_step   = std::numeric_limits<int>::max();
+    // maximum number of steps and stop time
+    int max_step = std::numeric_limits<int>::max();
     amrex::Real stop_time = std::numeric_limits<amrex::Real>::max();
 
+    // advective cfl number - dt = cfl*dx/umax
     amrex::Real cfl = 0.7;
 
+    // how often each level regrids the higher levels of refinement
+    // (after a level advances that many time steps)
     int regrid_int = 2;
 
+    // hyperbolic refluxing as part of multilevel synchronization
     int do_reflux = 1;
 
+    // plotfile prefix and frequency
     std::string plot_file {"plt"};
     int plot_int = -1;
 };
