--- Tutorials/Amr/Advection_AmrCore/Source/AmrAdvInit.cpp
@@ -1,51 +1,0 @@
-
-#include <AmrAdv.H>
-#include <AmrAdv_F.H>
-
-using namespace amrex;
-
-void
-AmrAdv::InitData ()
-{
-    const Real time = 0.0;
-    InitFromScratch(time);
-    AverageDown();
-
-    if (plot_int > 0) {
-        WritePlotFile();
-    }
-}
-
-void AmrAdv::MakeNewLevelFromScratch (int lev, Real time, const BoxArray& ba,
-				      const DistributionMapping& dm)
-{
-    const int ncomp = 1;
-    const int nghost = 0;
-
-    phi_new[lev].reset(new MultiFab(ba, dm, ncomp, nghost));
-    phi_old[lev].reset(new MultiFab(ba, dm, ncomp, nghost));
-
-    t_new[lev] = time;
-    t_old[lev] = time - 1.e200;
-
-    if (lev > 0 && do_reflux) {
-	flux_reg[lev].reset(new FluxRegister(ba, dm, refRatio(lev-1), lev, ncomp));
-    }
-
-    const Real* dx = geom[lev].CellSize();
-    const Real* prob_lo = geom[lev].ProbLo();
-    Real cur_time = t_new[lev];
-
-    MultiFab& state = *phi_new[lev];
-
-    for (MFIter mfi(state); mfi.isValid(); ++mfi)
-    {
-        const Box& box = mfi.validbox();
-        const int* lo  = box.loVect();
-        const int* hi  = box.hiVect();
-
-	initdata(lev, cur_time, ARLIM_3D(lo), ARLIM_3D(hi),
-		 BL_TO_FORTRAN_3D(state[mfi]), ZFILL(dx),
-		 ZFILL(prob_lo));
-    }
-}
